Extract AttrSignatureInfo
