+++ near-bindgen-core/src/lib.rs
@@ -1,6 +1,6 @@
 #![recursion_limit = "128"]
 use crate::code_generator::{method_wrapper, processed_impl_method};
-use crate::info_extractor::MethodInfo;
+use crate::info_extractor::ImplMethodInfo;
 use syn::export::TokenStream2;
 use syn::spanned::Spanned;
 use syn::{Error, ImplItem, ItemImpl};
@@ -26,7 +26,8 @@ pub fn process_impl(item_impl: &mut ItemImpl) -> TokenStream2 {
     let impl_type = item_impl.self_ty.as_ref();
     for subitem in item_impl.items.iter_mut() {
         if let ImplItem::Method(m) = subitem {
-            let method_info = MethodInfo::new((*m).clone(), (*impl_type).clone(), is_trait_impl);
+            let method_info =
+                ImplMethodInfo::new((*m).clone(), (*impl_type).clone(), is_trait_impl);
             let method_info = match method_info {
                 Ok(x) => x,
                 Err(err) => {
@@ -48,7 +49,7 @@ pub fn process_impl(item_impl: &mut ItemImpl) -> TokenStream2 {
 #[cfg(test)]
 mod tests {
     use syn::{Type, ImplItemMethod, parse_quote};
-    use crate::info_extractor::MethodInfo;
+    use crate::info_extractor::ImplMethodInfo;
     use crate::code_generator::method_wrapper;
     use quote::quote;
 
@@ -57,7 +58,7 @@ mod tests {
     fn trait_implt() {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod = syn::parse_str("fn method(&self) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -75,7 +76,7 @@ mod tests {
     fn no_args_no_return_no_mut() {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod = syn::parse_str("pub fn method(&self) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -93,7 +94,7 @@ mod tests {
     fn no_args_no_return_mut() {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod = syn::parse_str("pub fn method(&mut self) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -112,7 +113,7 @@ mod tests {
     fn arg_no_return_no_mut() {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod = syn::parse_str("pub fn method(&self, k: u64) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -141,7 +142,7 @@ mod tests {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod =
             syn::parse_str("pub fn method(&mut self, k: u64, m: Bar) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
                 #[cfg(target_arch = "wasm32")]
@@ -170,7 +171,7 @@ mod tests {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod =
             syn::parse_str("pub fn method(&mut self, k: u64, m: Bar) -> Option<u64> { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
                 #[cfg(target_arch = "wasm32")]
@@ -202,7 +203,7 @@ mod tests {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod =
             syn::parse_str("pub fn method(&self) -> &Option<u64> { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -223,7 +224,7 @@ mod tests {
     fn arg_ref() {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod = syn::parse_str("pub fn method(&self, k: &u64) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
                 #[cfg(target_arch = "wasm32")]
@@ -250,7 +251,7 @@ mod tests {
         let impl_type: Type = syn::parse_str("Hello").unwrap();
         let method: ImplItemMethod =
             syn::parse_str("pub fn method(&self, k: &mut u64) { }").unwrap();
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -278,7 +279,7 @@ mod tests {
         let method: ImplItemMethod = parse_quote! {
             pub fn method(&self, #[callback] x: &mut u64, y: String, #[callback] z: Vec<u8>) { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -318,7 +319,7 @@ mod tests {
         let method: ImplItemMethod = parse_quote! {
             pub fn method(&self, #[callback] x: &mut u64, #[callback] y: String) { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -351,7 +352,7 @@ mod tests {
         let method: ImplItemMethod = parse_quote! {
             pub fn method(&self, #[callback_vec] x: Vec<String>, y: String) { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -389,7 +390,7 @@ mod tests {
             #[init]
             pub fn method(k: &mut u64) -> Self { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -418,7 +419,7 @@ mod tests {
             #[result_serializer(borsh)]
             pub fn method(&mut self, #[serializer(borsh)] k: u64, #[serializer(borsh)]m: Bar) -> Option<u64> { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
@@ -451,7 +452,7 @@ mod tests {
         let method: ImplItemMethod = parse_quote! {
             pub fn method(&self, #[callback] #[serializer(borsh)] x: &mut u64, #[serializer(borsh)] y: String, #[callback] #[serializer(json)] z: Vec<u8>) { }
         };
-        let method_info = MethodInfo::new(method, impl_type, false).unwrap();
+        let method_info = ImplMethodInfo::new(method, impl_type, false).unwrap();
         let actual = method_wrapper(&method_info);
         let expected = quote!(
             #[cfg(target_arch = "wasm32")]
